Fix broken input file reduction across passes
With 8449a4270f93ee906e0519f0dcc5dd5c412f3d43 we broken the behavior
that a reduction from a pass was passed to the same file; instead only
reductions from the last pass had an effect.

This not only led to much increased runtime, but possibly also changed
output.